# docs/swagger.yaml
definitions:
...
  dto.Pagination:
    properties:
      current:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  echox.Response:
    properties:
      data: {}
      message: {}
    type: object
...
  models.User:
    properties:
      created_at:
        $ref: '#/definitions/sql.NullTime'
      created_by:
        type: string
      deleted:
        type: boolean
      email:
        type: string
      id:
        type: string
      password:
        type: string
      phone:
        type: string
      realname:
        type: string
      status:
        maximum: 1
        minimum: -1
        type: integer
      updated_at:
        $ref: '#/definitions/sql.NullTime'
      user_roles:
        items:
          $ref: '#/definitions/models.UserRole'
        type: array
      username:
        type: string
    required:
    - realname
    - status
    - username
    type: object
  models.UserQueryResult:
    properties:
      list:
        items:
          $ref: '#/definitions/models.User'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  models.UserRole:
    properties:
      created_at:
        $ref: '#/definitions/sql.NullTime'
      deleted:
        type: boolean
      id:
        type: string
      role_id:
        type: string
      updated_at:
        $ref: '#/definitions/sql.NullTime'
      user_id:
        type: string
    type: object
  sql.NullTime:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /api/v1/users:
    get:
      parameters:
      - in: query
        name: current
        type: integer
      - enum:
        - ASC
        - DESC
        in: query
        name: direction
        type: string
        x-enum-varnames:
        - OrderByASC
        - OrderByDESC
      - in: query
        name: key
        type: string
      - in: query
        maximum: 128
        name: pageSize
        type: integer
      - in: query
        name: queryPassword
        type: boolean
      - in: query
        name: queryValue
        type: string
      - in: query
        name: realname
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: roleIDs
        type: array
      - in: query
        maximum: 1
        minimum: -1
        name: status
        type: integer
      - in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.UserQueryResult'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: User Query
      tags:
      - User
    post:
      parameters:
      - description: User
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: User Create
      tags:
      - User
  /api/v1/users/{id}:
    delete:
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: User Delete By ID
      tags:
      - User
    get:
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: User Get By ID
      tags:
      - User
    put:
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      - description: User
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: User Update By ID
      tags:
      - User
  /api/v1/users/{id}/disable:
    put:
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: User Disable By ID
      tags:
      - User
  /api/v1/users/{id}/enable:
    put:
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echox.Response'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echox.Response'
      summary: User Enable By ID
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"